+++ examples/Products/get-products.php
@@ -3,7 +3,7 @@
 use Printful\Exceptions\PrintfulApiException;
 use Printful\Exceptions\PrintfulException;
 use Printful\PrintfulApiClient;
-use Printful\PrintfulProductsApi;
+use Printful\PrintfulProducts;
 use Printful\Structures\Sync\Responses\SyncProductsResponse;
 
 require_once __DIR__ . '/../../vendor/autoload.php';
@@ -22,7 +22,7 @@ try {
     $pf = new PrintfulApiClient($apiKey);
 
     // create Products Api object
-    $productsApi = new PrintfulProductsApi($pf);
+    $productsApi = new PrintfulProducts($pf);
 
     // set some paging info
     $offset = 0;
