+++ examples/Products/product-creation-oop.php
@@ -4,7 +4,7 @@ use Printful\Exceptions\PrintfulApiException;
 use Printful\Exceptions\PrintfulException;
 use Printful\Exceptions\PrintfulSdkException;
 use Printful\PrintfulApiClient;
-use Printful\PrintfulProductsApi;
+use Printful\PrintfulProducts;
 use Printful\Structures\Sync\Requests\SyncProductRequest;
 use Printful\Structures\Sync\Requests\SyncVariantRequest;
 use Printful\Structures\Sync\Requests\SyncVariantRequestFile;
@@ -26,7 +26,7 @@ try {
     $pf = new PrintfulApiClient($apiKey);
 
     // create Products Api object
-    $productsApi = new PrintfulProductsApi($pf);
+    $productsApi = new PrintfulProducts($pf);
 
     // create product request
     $productRequest = new SyncProductRequest;
